backend/requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
pymongo==4.15.3
python-dotenv==1.1.1
pydantic==2.12.3
pydantic-settings==2.11.0
motor==3.7.1
